Implement a basic Y-distortion, with squishy circles
Gosh this looks *neat.* The final distortion is happening with a
minimum resolution of 2px... I dunno strips? Scanlines per section,
whatever. Because the sine wave occasionally causes the updates to
get ahead of the scanning beam (reversing the image) this causes
some minor artifacts, but they end up looking more like interlaced
television and we can pretend they're on purpose. The intended
"lava lamp" effect reads just fine.

Super happy with this!

We discovered this is NOT working on PowerPak for unclear reasons
that need additional investigation.