Initialize a sprite table and do the OAM DMA thing
I didn't notice this until I tested on real hardware, but duh, OAM
isn't initialized to anything but random data, and that means on
real hardware we get fun random sprites littering the screen. Until
we plan to use sprites (maybe not at all in this demo?) set all of
those to offscreen.