Read in the basics of FM and SSG instruments
Note that for the moment, I am choosing to ignore most FM sequences, largely because I do not believe the 6502 has the processing power to deal with them in a useful way. I haven't run into any modules in the wild that use them, seems to be a bit of an advanced technique.